Norminette

+++ ft_printf/display_utils/ft_putnbr_fd.c
@@ -6,7 +6,7 @@
 /*   By: mlabouri <mlabouri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/11 17:48:10 by mlabouri          #+#    #+#             */
-/*   Updated: 2020/02/07 10:17:28 by mlabouri         ###   ########.fr       */
+/*   Updated: 2020/02/10 09:35:10 by mlabouri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putnbr_fd(int n, int fd)
 {
-	long int nb;
+	long int	nb;
 
 	if (fd < 0)
 		return ;
@@ -24,9 +24,9 @@ void	ft_putnbr_fd(int n, int fd)
 		nb = -nb;
 		ft_putchar_fd('-', 1);
 	}
-		if (nb > 9)
+	if (nb > 9)
 	{
-		ft_putnbr_fd(nb / 10, fd);
+		ft_putnbr_fd((int)nb / 10, fd);
 		nb = nb % 10;
 	}
 	nb = nb + 48;
